httphelper class implement
